+++ app/src/main/java/chat/rocket/android/helper/MessageParser.kt
@@ -20,6 +20,7 @@ import chat.rocket.android.R
 import chat.rocket.android.chatroom.viewmodel.MessageViewModel
 import chat.rocket.android.customtab.CustomTab
 import chat.rocket.android.customtab.WebViewFallback
+import chat.rocket.android.helper.ToastHelper
 import chat.rocket.android.widget.emoji.EmojiParser
 import chat.rocket.android.widget.emoji.EmojiRepository
 import chat.rocket.android.widget.emoji.EmojiTypefaceSpan
@@ -32,9 +33,8 @@ import ru.noties.markwon.SpannableConfiguration
 import ru.noties.markwon.renderer.SpannableMarkdownVisitor
 import java.util.regex.Pattern
 import javax.inject.Inject
-import android.widget.Toast
 import android.widget.TextView
-import timber.log.Timber
+import android.widget.Toast
 
 class MessageParser @Inject constructor(val context: Application, private val configuration: SpannableConfiguration) {
 
@@ -176,14 +176,9 @@ class MessageParser @Inject constructor(val context: Application, private val co
                         override fun onClick(view: View) {
                             with (view) {
                                 if (link.startsWith("http://www.requiresinternet.com")) {
-                                    val layout = LayoutInflater.from(context).inflate(R.layout.custom_toast, findViewById(R.id.custom_toast_container))
-                                    val text = layout.findViewById<TextView>(R.id.text)
-                                    text.setText("Can I help you get access to this content?")
-                                    val toast =  Toast(context)
-                                    toast.setGravity(Gravity.FILL_HORIZONTAL or Gravity.BOTTOM, 0, 0)
-                                    toast.setDuration(Toast.LENGTH_LONG)
-                                    toast.setView(layout)
-                                    toast.show()
+                                    ToastHelper.showCustomToast(context,
+                                            "Can I help you get access to this content?",
+                                            this)
                                 } else {
                                     CustomTab.openCustomTab(context, link, WebViewFallback())
                                 }
