+++ app/src/main/res/layout/custom_toast.xml
@@ -17,5 +17,5 @@
         android:layout_height="wrap_content"
         android:gravity="left"
         android:textColor="#FFFFFF"
-        android:textSize="20dp" />
+        android:textSize="24dp" />
 </LinearLayout>
